+++ VisualPinball.Unity/VisualPinball.Unity.Editor/VPT/HitTarget/TargetColliderInspector.cs
@@ -34,6 +34,7 @@ namespace VisualPinball.Unity.Editor
 		private SerializedProperty _elasticityProperty;
 		private SerializedProperty _elasticityFalloffProperty;
 		private SerializedProperty _frictionProperty;
+		private SerializedProperty _rollingResistanceProperty;
 		private SerializedProperty _scatterProperty;
 
 		protected override void OnEnable()
@@ -49,6 +50,8 @@ namespace VisualPinball.Unity.Editor
 			_elasticityProperty = serializedObject.FindProperty(nameof(HitTargetColliderComponent.Elasticity));
 			_elasticityFalloffProperty = serializedObject.FindProperty(nameof(HitTargetColliderComponent.ElasticityFalloff));
 			_frictionProperty = serializedObject.FindProperty(nameof(HitTargetColliderComponent.Friction));
+			_rollingResistanceProperty = serializedObject.FindProperty(
+				nameof(HitTargetColliderComponent.RollingResistance));
 			_scatterProperty = serializedObject.FindProperty(nameof(HitTargetColliderComponent.Scatter));
 		}
 
@@ -88,6 +91,7 @@ namespace VisualPinball.Unity.Editor
 				PropertyField(_elasticityProperty);
 				PropertyField(_elasticityFalloffProperty);
 				PropertyField(_frictionProperty);
+				PropertyField(_rollingResistanceProperty);
 				PropertyField(_scatterProperty, "Scatter Angle");
 				EditorGUI.EndDisabledGroup();
 			}