+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Spinner/SpinnerColliderComponent.cs
@@ -69,6 +69,11 @@ namespace VisualPinball.Unity
 			set { }
 		}
 
+		public override float PhysicsRollingResistance {
+			get => 0;
+			set { }
+		}
+
 		public override float PhysicsScatter {
 			get => 0;
 			set { }